Chuyển đổi các entity có id Long sang UUID

+++ MobileAPI-develop/src/main/java/com/example/mobileapi/dto/response/MonthlyRevenueResponse.java
@@ -3,12 +3,14 @@ package com.example.mobileapi.dto.response;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Builder
 @Getter
 public class MonthlyRevenueResponse {
     int month;
-    long revenue;
+    BigDecimal revenue;
 
 
 }
